requirements.txt
azure-cosmos
requests
